third_party/WebKit/LayoutTests/fast/xpath/invalid-resolver.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="console"></div>

<script>
    shouldThrow("document.evaluate('/body', document, 5, 0, null)", '"TypeError: Failed to execute \'evaluate\' on \'Document\': The 3rd argument provided is either null, or an invalid XPathNSResolver object."');
    shouldThrow("document.evaluate('/body', document, 'a resolver', 0, null)", '"TypeError: Failed to execute \'evaluate\' on \'Document\': The 3rd argument provided is either null, or an invalid XPathNSResolver object."');
    shouldBe("document.evaluate('/body', document, Date, 0, null).toString()", '"[object XPathResult]"');
</script>
</body>
</html>
