third_party/WebKit/LayoutTests/fast/xpath/invalid-resolver-expected.txt

PASS document.evaluate('/body', document, 5, 0, null) threw exception TypeError: Failed to execute 'evaluate' on 'Document': The 3rd argument provided is either null, or an invalid XPathNSResolver object..
PASS document.evaluate('/body', document, 'a resolver', 0, null) threw exception TypeError: Failed to execute 'evaluate' on 'Document': The 3rd argument provided is either null, or an invalid XPathNSResolver object..
PASS document.evaluate('/body', document, Date, 0, null).toString() is "[object XPathResult]"
PASS successfullyParsed is true

TEST COMPLETE
